canonicalRequest: Allow "UNSIGNED-PAYLOAD"
In the `canonicalRequest()` the payload is always hashed. But when using `AWS S3 GetObject` we need to provide the payload as is: "UNSIGNED-PAYLOAD". Otherwise our signing does not match.

https://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-query-string-auth.html
```
You don't include a payload hash in the Canonical Request, because when you create a presigned URL, you don't know the payload content because the URL is used to upload an arbitrary payload. Instead, you use a constant string UNSIGNED-PAYLOAD.
```